src/workloadgenerator/requirements.txt
azure-storage-blob==12.8.1
Flask==2.0.1
gunicorn==20.1.0